+++ src/main/php/lang/Thread.class.php
@@ -163,9 +163,12 @@ class Thread extends Object {
       $this->_pid= $parent;
       try {
         exit($this->getTarget()->run());
-      } catch (\SystemExit $e) {
+      } catch (SystemExit $e) {
         if ($message= $e->getMessage()) echo $message, "\n";
         exit($e->getCode());
+      } catch (\Exception $e) {
+        fputs(STDERR, 'Uncaught exception (in child #'.$this->_id.'): '.\xp::stringOf($e));
+        exit(0xf0);
       } catch (\Throwable $t) {
         fputs(STDERR, 'Uncaught exception (in child #'.$this->_id.'): '.\xp::stringOf($t));
         exit(0xf0);