+++ src/main/php/util/PropertyManager.class.php
@@ -83,7 +83,7 @@ class PropertyManager extends \lang\Object {
       foreach ($sources as $source) {
         $this->appendSource($source);
       }
-    } catch (\lang\IllegalArgumentException $e) {
+    } catch (\Exception $e) {
       $this->provider= $provider;
       throw $e;
     } catch (\Throwable $e) {